restructure and document import export plugin
